requirements.txt
flask>=3.0.0,<4
tensorflow-cpu==2.17.*
pillow>=10.0.0
numpy>=1.24.0,<2.0
